End SupposeStream output when no expectations
If there are no expectations, the output stream can be ended immediately
(as it is after the last expectation is met).  Check for this condition
when the expect Array is empty after being dequeued.  Update the test
suite to test for this condition.

This also fixes a bug where the output stream would not be ended if the
last expectation didn't have a response (or null pushed multiple times
for multiple responses).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>